Add missing headers for string functions and fix printfs for cd_nelmts (#60)
* add missing headers for `strncasecmp` and `strcasestr` for Ubuntu (where actions are run)
* fix printf args for `cd_nelmts`

+++ test/test_write.c
@@ -323,8 +323,8 @@ static hid_t setup_filter(int n, hsize_t *chunk, int zfpmode,
         cd_nelmts = 0; /* causes default behavior of ZFP library */
 
     /* print cd-values array used for filter */
-    printf("%d cd_values= ",cd_nelmts);
-    for (i = 0; i < cd_nelmts; i++)
+    printf("%d cd_values= ", (int) cd_nelmts);
+    for (i = 0; i < (int) cd_nelmts; i++)
         printf("%u,", cd_values[i]);
     printf("\n");
